refactor(telos): semantic folders + DRY action application
Flat src/ → folders mirroring the package's own design:

  core/        the Aristotelian engine, split by concept (mover, ledger, domain,
               events, deliberator, eudaimon) behind a barrel
  deliberators/  rule, ai
  socratic/    daimonion (veto), elenchus
  platonic/    demiurge

Also: extracted the duplicated apply→record / veto→ctx.vetoed flow out of both
deliberators into a single core applyAction(); moved VetoError into core so the
layering is one-way (core never depends on the socratic/platonic extensions).

No behavior change — 20 tests, tsc, lint, knip, example all green.
