+++ packages/telos/package.json
@@ -14,10 +14,10 @@
   },
   "exports": {
     ".": "./src/index.ts",
-    "./ai": "./src/deliberate-ai.ts",
-    "./daimonion": "./src/daimonion.ts",
-    "./elenchus": "./src/elenchus.ts",
-    "./demiurge": "./src/demiurge.ts"
+    "./ai": "./src/deliberators/ai.ts",
+    "./daimonion": "./src/socratic/daimonion.ts",
+    "./elenchus": "./src/socratic/elenchus.ts",
+    "./demiurge": "./src/platonic/demiurge.ts"
   },
   "dependencies": {
     "zod": "^4.0.0"
