+++ packages/telos/src/core/eudaimon.ts
@@ -1,104 +1,8 @@
-// Domain-agnostic core. Depends only on zod — never the AI SDK, so the offline
-// path stays dependency-light (the LLM deliberator lives in ./deliberate-ai).
-
-import type { z } from "zod";
-
-export type Awaitable<T> = T | Promise<T>;
-
-// Where a goal version came from. The anchor (authority) is the fixed parent
-// telos; engine versions are subordinate goals the agent proposed in service of it.
-export type GoalSource = "authority" | "engine";
-
-// The goal: frozen, behaviorless target data. It moves the agent only by being
-// the thing measured against — it never acts and is never mutated.
-export class UnmovedMover<T> {
-  readonly tenant: string;
-  readonly version: number;
-  readonly target: T;
-  readonly source: GoalSource;
-
-  constructor(init: {
-    tenant: string;
-    version: number;
-    target: T;
-    source?: GoalSource;
-  }) {
-    this.tenant = init.tenant;
-    this.version = init.version;
-    this.target = init.target;
-    this.source = init.source ?? "authority";
-    Object.freeze(this);
-  }
-}
-
-// Append-only history of goals; installing a new version is the only way a goal
-// changes. Methods are Awaitable so a DB-backed ledger drops in unchanged.
-export interface GoalLedger<T> {
-  install(
-    tenant: string,
-    target: T,
-    source?: GoalSource,
-  ): Awaitable<UnmovedMover<T>>;
-  // The current working goal (latest of any source). What the agent pursues.
-  // Named `latest`, not `current`, to dodge the repo's .current lint rule.
-  latest(tenant: string): Awaitable<UnmovedMover<T>>;
-  // The fixed parent telos: latest AUTHORITY-installed version. The engine never
-  // overwrites it — that immovability is the safety spine of anchored proposing.
-  anchor(tenant: string): Awaitable<UnmovedMover<T>>;
-  history(tenant: string): Awaitable<readonly UnmovedMover<T>[]>;
-}
-
-export interface PursuitContext {
-  readonly tenant: string;
-  readonly moverVersion: number;
-  record(kind: string, payload?: unknown): Promise<void>;
-  // The daimonion forbade this action; it was not applied. (See ./daimonion.)
-  vetoed(kind: string, reason: string, payload?: unknown): Promise<void>;
-}
-
-// A framework-agnostic "hand": the rule deliberator calls it via plan(), the AI
-// deliberator wraps it as an LLM tool. Write it once; both paths use it.
-export interface Action<P = unknown> {
-  kind: string;
-  description: string;
-  schema: z.ZodType<P>;
-  apply(tenant: string, input: P): Promise<void>;
-}
-
-export interface Domain<S, T, G = unknown> {
-  readonly name: string;
-  observe(tenant: string): Promise<S>;
-  satisfied(state: S, target: T): boolean;
-  gap(state: S, target: T): G;
-  readonly instructions: string;
-  readonly actions: Action[];
-  prompt(input: {
-    state: S;
-    target: T;
-    gap: G;
-    tenant: string;
-    moverVersion: number;
-  }): string;
-  // Deterministic steps so the rule deliberator can run without an LLM.
-  plan?(input: {
-    state: S;
-    target: T;
-    gap: G;
-  }): Array<{ kind: string; input: unknown }>;
-}
-
-// The reasoning engine — swap rule-based <-> AI without touching the core.
-export interface Deliberator {
-  run<S, T, G>(args: {
-    domain: Domain<S, T, G>;
-    state: S;
-    target: T;
-    gap: G;
-    ctx: PursuitContext;
-    instructions: string;
-    prompt: string;
-  }): Promise<{ summary: string; actionsTaken: string[] }>;
-}
+import type { Awaitable } from "./awaitable";
+import type { Deliberator } from "./deliberator";
+import type { Domain, PursuitContext } from "./domain";
+import type { EventBus } from "./events";
+import type { GoalLedger } from "./ledger";
 
 // Anchored self-direction: after a cycle the engine may author the next
 // SUBORDINATE goal — but `anchor` (the fixed parent) is handed in so the proposal
@@ -117,54 +21,6 @@ export type ApproveGoal<T> = (
   proposed: T,
   ctx: { tenant: string },
 ) => Awaitable<boolean>;
-
-export type DomainEvent<T> =
-  | { type: "state.changed"; tenant: string }
-  | { type: "goal.updated"; tenant: string; target: T }
-  | {
-      type: "eudaimon.action.applied";
-      tenant: string;
-      moverVersion: number;
-      kind: string;
-      payload?: unknown;
-    }
-  // The Eudaimon's action, forbidden by a Daimonion before it ran (see ./daimonion).
-  | {
-      type: "eudaimon.action.vetoed";
-      tenant: string;
-      moverVersion: number;
-      kind: string;
-      reason: string;
-      payload?: unknown;
-    }
-  | {
-      type: "eudaimon.pursued";
-      tenant: string;
-      moverVersion: number;
-      summary: string;
-    }
-  | { type: "unmovedMover.reached"; tenant: string; moverVersion: number }
-  | {
-      type: "eudaimon.goal.proposed";
-      tenant: string;
-      moverVersion: number;
-      target: T;
-    }
-  | { type: "eudaimon.goal.rejected"; tenant: string; target: T };
-
-export type DomainEventType<T> = DomainEvent<T>["type"];
-
-export type EventHandler<T, K extends DomainEventType<T>> = (
-  event: Extract<DomainEvent<T>, { type: K }>,
-) => Promise<void>;
-
-export interface EventBus<T> {
-  publish(event: DomainEvent<T>): Promise<void>;
-  subscribe<K extends DomainEventType<T>>(
-    type: K,
-    handler: EventHandler<T, K>,
-  ): void;
-}
 
 export interface EudaimonDeps<S, T, G = unknown> {
   tenant: string;
