+++ packages/telos/src/socratic/daimonion.test.ts
@@ -1,9 +1,12 @@
 import { describe, expect, test } from "bun:test";
 import { z } from "zod";
-import { type Domain, Eudaimon } from "./core";
-import { inMemoryBus } from "./bus";
-import { InMemoryGoalLedger } from "./ledger";
-import { ruleDeliberator } from "./deliberate-rule";
+import {
+  type Domain,
+  Eudaimon,
+  inMemoryBus,
+  InMemoryGoalLedger,
+} from "../core";
+import { ruleDeliberator } from "../deliberators/rule";
 import { type Daimonion, guardedBy } from "./daimonion";
 
 interface CounterState {
