+++ packages/telos/src/socratic/elenchus.ts
